sharepoint-ai-agent/requirements.txt
openai==1.50.2
httpx==0.27.2
streamlit==1.48.1
Office365-REST-Python-Client
pypdf
pdf2image
pillow
pytesseract
pinecone
langextract
tiktoken
python-dotenv
